+++ .circleci/config.yml
@@ -61,7 +61,7 @@ jobs:
       # Run tests and build the binaries.
       - run:
           name: Run tests
-          command: go test -cover -v $(go list ./... | circleci tests split)
+          command: go test -cover -tags integration -v $(go list ./... | circleci tests split)
       - run:
           name: Build binary
           command: go build -o ./build/_output/bin/dynatrace-oneagent-operator ./cmd/manager
