+++ pkg/controller/oneagent/oneagent_controller_integration_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package oneagent
 
 import (
